Move Content domain classes to `content` namespace
