+++ app/domain/content/importers/all_inventory.rb
@@ -1,5 +1,5 @@
-module Importers
-  class AllInventory
+module Content
+  class Importers::AllInventory
     def run
       Importers::AllContentItems.new.run
     end
