+++ spec/domain/content/importers/all_inventory_spec.rb
@@ -1,0 +1,11 @@
+module Content
+  RSpec.describe Importers::AllInventory do
+    describe '#run' do
+      it 'imports all the inventory' do
+        expect_any_instance_of(Importers::AllContentItems).to receive(:run)
+
+        Importers::AllInventory.new.run
+      end
+    end
+  end
+end
